miner: fix/rename miner actor withdraw
