+++ cmd/lotus-storage-miner/actor.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	"fmt"
+	"github.com/filecoin-project/specs-actors/actors/builtin"
+	"golang.org/x/xerrors"
 
 	ma "github.com/multiformats/go-multiaddr"
 	"github.com/urfave/cli/v2"
@@ -19,6 +21,7 @@ var actorCmd = &cli.Command{
 	Usage: "manipulate the miner actor",
 	Subcommands: []*cli.Command{
 		actorSetAddrsCmd,
+		actorWithdrawCmd,
 	},
 }
 
@@ -89,5 +92,78 @@ var actorSetAddrsCmd = &cli.Command{
 		fmt.Printf("Requested multiaddrs change in message %s\n", smsg.Cid())
 		return nil
 
+	},
+}
+
+
+var actorWithdrawCmd = &cli.Command{
+	Name:  "withdraw",
+	Usage: "withdraw available balance",
+	ArgsUsage: "[amount (FIL)]",
+	Action: func(cctx *cli.Context) error {
+		nodeApi, closer, err := lcli.GetStorageMinerAPI(cctx)
+		if err != nil {
+			return err
+		}
+		defer closer()
+
+		api, acloser, err := lcli.GetFullNodeAPI(cctx)
+		if err != nil {
+			return err
+		}
+		defer acloser()
+
+		ctx := lcli.ReqContext(cctx)
+
+		maddr, err := nodeApi.ActorAddress(ctx)
+		if err != nil {
+			return err
+		}
+
+		mi, err := api.StateMinerInfo(ctx, maddr, types.EmptyTSK)
+		if err != nil {
+			return err
+		}
+
+		available, err := api.StateMinerAvailableBalance(ctx, maddr, types.EmptyTSK)
+		if err != nil {
+			return err
+		}
+
+		amount := available
+		if cctx.Args().Present() {
+			f, err := types.ParseFIL(cctx.Args().First())
+			if err != nil {
+				return xerrors.Errorf("parsing 'amount' argument: %w", err)
+			}
+
+			amount = abi.TokenAmount(f)
+
+			if amount.GreaterThan(available) {
+				return xerrors.Errorf("can't withdraw more funds than available; requested: %s; available: %s", amount, available)
+			}
+		}
+
+		params, err := actors.SerializeParams(&miner.WithdrawBalanceParams{
+			AmountRequested: amount, // Default to attempting to withdraw all the extra funds in the miner actor
+		})
+		if err != nil {
+			return err
+		}
+
+		smsg, err := api.MpoolPushMessage(ctx, &types.Message{
+			To:     maddr,
+			From:   mi.Owner,
+			Value:  types.NewInt(0),
+			Method: builtin.MethodsMiner.WithdrawBalance,
+			Params: params,
+		}, nil)
+		if err != nil {
+			return err
+		}
+
+		fmt.Printf("Requested rewards withdrawal in message %s\n", smsg.Cid())
+
+		return nil
 	},
 }
